Adding tests for collectBond
